Module engine_test::ctor

Create a storage engine with a concrete type. This should ultimately be the only module within TiKv that needs to know about concrete engines. Other code only uses the engine_traits abstractions.

At the moment this has a lot of open-coding of engine-specific initialization, but in the future more constructor abstractions should be pushed down into engine_traits.

This module itself is intended to be extracted from this crate into its own crate, once the requirements for engine construction are better understood.

Modules

panic

rocks

Structs

CFOptions

ColumnFamilyOptions
Properties for a single column family

DBOptions

Enums

CryptoOptions

Traits

EngineConstructorExt
Engine construction