Trait engine_test::ctor::EngineConstructorExt

pub trait EngineConstructorExt: Sized {
    fn new_engine(
        path: &str, 
        db_opt: Option<DBOptions>, 
        cfs: &[&str], 
        opts: Option<Vec<CFOptions<'_>>>
    ) -> Result<Self>;

    fn new_engine_opt(
        path: &str, 
        db_opt: DBOptions, 
        cfs_opts: Vec<CFOptions<'_>>
    ) -> Result<Self>;
}

Engine construction

For simplicity, all engine constructors are expected to configure every engine such that all of TiKV and its tests work correctly, for the constructed column families.

Specifically, this means that RocksDB constructors should set up all properties collectors, always.

Required methods

fn new_engine(
path: &str,
db_opt: Option<DBOptions>,
cfs: &[&str],
opts: Option<Vec<CFOptions<'_>>>
) -> Result<Self>

Create a new engine with either:

The column families specified as cfs, with default options, or

The column families specified as opts, with options.

Note that if opts is not None then the cfs argument is completely ignored.

fn new_engine_opt(
path: &str,
db_opt: DBOptions,
cfs_opts: Vec<CFOptions<'_>>
) -> Result<Self>

Create a new engine with specified column families and options

Loading content...